Added function support.
Functions are added to my lisp interpreter.
The currently implemented functions do not suppord parameters nor return values. Functions can be executed, if a list contains a single element which represents the name of the function.
Furthermore, the first function to be executed is the "main"-function, which mustn't have any parameters.